+++ cmd/singleplayer/main.go
@@ -2,12 +2,11 @@ package main
 
 import (
 	"flag"
-	"fmt"
 	"log"
 	"math/rand"
-	"sort"
 
-	pd "github.com/jcpage573/prisoners-dilemma/src"
+	"github.com/jcpage573/prisoners-dilemma/internal/game"
+	"github.com/jcpage573/prisoners-dilemma/internal/prisoner"
 )
 
 // Example Strategies
@@ -56,8 +55,8 @@ func megatron(oa []bool, opa []bool) bool {
 	return rand.Intn(2) == 1
 }
 
-func CreatePrisoners() []pd.Prisoner {
-	prisoners := []pd.Prisoner{
+func CreatePrisoners() []prisoner.Prisoner {
+	prisoners := []prisoner.Prisoner{
 		// Jackson
 		{Name: "Random", Owner: "Jackson", Strategy: random},
 		{Name: "Mod 3 Prisoner", Owner: "Jackson", Strategy: defectCooperateTwiceThenDefect},
@@ -82,38 +81,6 @@ func CreatePrisoners() []pd.Prisoner {
 	return prisoners
 }
 
-func prettyPrintMap(m map[string]int, title string) {
-	// Extract keys and sort them
-	keys := make([]string, 0, len(m))
-	for key := range m {
-		keys = append(keys, key)
-	}
-	sort.Strings(keys)
-
-	// Print the map in a pretty format
-	fmt.Printf("%s:\n", title)
-	fmt.Println("---------------")
-	for _, key := range keys {
-		fmt.Printf("%-15s : %d\n", key, m[key])
-	}
-}
-
-// Function to calculate and print average scores
-func printAverageScores(totals map[string]int, counts map[string]int, title string) {
-	keys := make([]string, 0, len(counts))
-	for key := range counts {
-		keys = append(keys, key)
-	}
-	sort.Strings(keys)
-
-	fmt.Printf("%s:\n", title)
-	fmt.Println("---------------")
-	for _, key := range keys {
-		average := float64(totals[key]) / float64(counts[key])
-		fmt.Printf("%-15s : %.2f\n", key, average)
-	}
-}
-
 func main() {
 	n := flag.Int("n", 200, "Number of games to play")
 	manual := flag.Bool("manual", false, "Play the game manually via CLI")
@@ -123,21 +90,7 @@ func main() {
 		log.Fatalln("NO MANUAL")
 		// manualPlay()
 	} else {
-		totals := make(map[string]int)
-		oTotals := make(map[string]int)
-		ownerCounts := make(map[string]int)
 		prisoners := CreatePrisoners()
-		for _, i := range pd.PlayGames(prisoners, *n) {
-			// fmt.Println(i.Results())
-			p1s, p2s := i.Scores()
-			totals[i.Player1.Name] += p1s
-			oTotals[i.Player1.Owner] += p1s
-			totals[i.Player2.Name] += p2s
-			oTotals[i.Player2.Owner] += p2s
-			ownerCounts[i.Player1.Owner]++
-			ownerCounts[i.Player2.Owner]++
-		}
-		prettyPrintMap(totals, "\nPlayer Scores")
-		printAverageScores(oTotals, ownerCounts, "\n\nOwner Average Scores")
+		game.Sim(prisoners, *n)
 	}
 }
